+++ src/main/java/org/antbear/tododont/backend/security/service/scheduling/RegistrationMailScheduler.java
@@ -8,6 +8,7 @@ import org.antbear.tododont.backend.security.service.SecurityMailSender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,42 +35,44 @@ public class RegistrationMailScheduler {
     @Autowired
     private SecurityMailSender mailSender;
 
-    private int maxAttempts = 10;
+    @Autowired
+    private ScheduleExecuteStrategy scheduleExecuteStrategy;
+
+    @Value("${app.mail.schedule.maxAttempts}")
+    private int maxAttempts;
 
     public int getMaxAttempts() {
-        return maxAttempts;
-    }
-
-    public void setMaxAttempts(final int maxAttempts) {
-        this.maxAttempts = maxAttempts;
+        return this.maxAttempts;
     }
 
     @Transactional
-    @Scheduled(fixedDelay = 2 * 60 * 1000) // TODO unrealistic, better use exponential back off
+    @Scheduled(fixedDelay = 2 * 60 * 1000)
     public void onSchedule() {
         boolean success;
         for (final SecurityTokenMailSchedule srm : this.registrationMailScheduleDao.findAll()) {
             log.debug("Processing scheduled registration mail {}", srm);
 
-            success = false;
-            try {
-                srm.setAttempts(1 + srm.getAttempts());
-                srm.setLastAttempt(new Date());
-                processRegistrationMail(srm);
-                success = true;
-            } catch (Exception ex) {
-                log.warn("Failed processing scheduled registration mail " + srm, ex);
-            } finally {
-                if (success) {
-                    log.info("Done sending registration to mail to {}", srm.getEmail());
-                    this.registrationMailScheduleDao.delete(srm.getPK());
-                } else {
-                    if (srm.getAttempts() < maxAttempts) {
-                        this.registrationMailScheduleDao.update(srm);
+            if (this.scheduleExecuteStrategy.isSchedulingAttemptDesired(srm.getAttempts(), srm.getLastAttempt())) {
+                success = false;
+                try {
+                    srm.setAttempts(1 + srm.getAttempts());
+                    srm.setLastAttempt(new Date());
+                    processRegistrationMail(srm);
+                    success = true;
+                } catch (Exception ex) {
+                    log.warn("Failed processing scheduled registration mail " + srm, ex);
+                } finally {
+                    if (success) {
+                        log.info("Done sending registration to mail to {}", srm.getEmail());
+                        this.registrationMailScheduleDao.delete(srm.getPK());
                     } else {
-                        log.warn("Max attempts of scheduled registration mail reached, deleting user {}", srm.getEmail());
-                        // scheduled task will be deleted by dropping user
-                        this.userDetailsService.deleteUser(srm.getEmail());
+                        if (srm.getAttempts() < maxAttempts) {
+                            this.registrationMailScheduleDao.update(srm);
+                        } else {
+                            log.warn("Max attempts of scheduled registration mail reached, deleting user {}", srm.getEmail());
+                            // scheduled task will be deleted by dropping user
+                            this.userDetailsService.deleteUser(srm.getEmail());
+                        }
                     }
                 }
             }
